dijkstra fixed + position fixed
