+++ project/project.pyde
@@ -2,21 +2,20 @@
 from Vehicle import Vehicle
 from Food import Food
 from Terrain import Terrain
-from Bfs import Bfs
-from Dijkstra import Dijkstra
+from Searchs import Bfs, Dijkstra
 from random import randint
 
 
 def getNonObstacle(mat, w, h):
-    pos = [randint(0, w), randint(0, h)]
+    pos = [randint(0, w/10), randint(0, h/10)]
     
-    while mat[int(floor(pos[0]/10))][int(floor(pos[1]/10))] == 0:
-        pos = [randint(0, w), randint(0, h)]
+    while mat[int(floor(pos[0]))][int(floor(pos[1]))] == 0:
+        pos = [randint(0, w/10), randint(0, h/10)]
         
-    return PVector(pos[0], pos[1])
+    return PVector(pos[0] * 10 - 5, pos[1] * 10 - 5)
 
 def setup():
-    global vehicle, food, counter, printer, terreno, currentState, path, finalPath, bfs
+    global vehicle, food, counter, printer, terreno, currentState, path, finalPath, search
     size(640, 360)
     background(0)
     finalPath = []
@@ -31,11 +30,9 @@ def setup():
     fPos = getNonObstacle(terreno.matrixL, width-10, height-10)
     vehicle = Vehicle(vPos, velocity_v)
     food = Food(fPos, velocity_f)
-    #bfs = Bfs(vehicle.position, food.getPosition(), terreno)
-    bfs = Dijkstra(vehicle.position, food.getPosition(), terreno)
-    #frameRate(5)
-    #noLoop()
-    #x = path.get()
+    #search = Bfs(vehicle.position, food.getPosition(), terreno)
+    search = Dijkstra(vehicle.position, food.getPosition(), terreno)
+    
 def draw(): 
     global currentState, pathi
     global counter
@@ -44,12 +41,12 @@ def draw():
     terreno.render()
     fill(255)
     if currentState == 0:
-        ret = bfs.djikstra_search()
-        for pathi in list(bfs.frontier.queue):
+        ret = search.djikstra_search()
+        for pathi in list(search.frontier):
             fill(170, 200, 255)
-            rect(pathi.x * 10, pathi.y * 10, 10, 10)
-        if bfs.frontier.empty() or ret == 1:
-            path = bfs.came_from
+            rect(pathi[1].x * 10, pathi[1].y * 10, 10, 10)
+        if search.frontier == [] or ret == 1:
+            path = search.came_from
             currentState = 1
     elif currentState == 1:    
         finalPos = PVector(floor(food.position[0]/10), floor(food.position[1]/10))
@@ -77,7 +74,7 @@ def draw():
             finalPath = []
             counter += 1
             currentState = 0
-            bfs.reset(vehicle.position, food.position, terreno)
+            search.reset(vehicle.position, food.position, terreno)
     
     vehicle.display()
     food.display()
